Fixed an issue related to flaot precision when computing iso values from MC grid.

+++ include/mesh.h
@@ -26,17 +26,17 @@ public:
     ~Mesh();
 
     std::vector<Triangle>& triangles() {return _triangles;}
-    std::vector<glm::dvec3>& points() {return _points;}
-    std::vector<glm::dvec3>& normals() {return _normals;}
+    std::vector<glm::vec3>& points() {return _points;}
+    std::vector<glm::vec3>& normals() {return _normals;}
 
     const std::vector<Triangle>& triangles() const  {return _triangles;}
-    const std::vector<glm::dvec3>& points() const {return _points;}
-    const std::vector<glm::dvec3>& normals() const {return _normals;}
+    const std::vector<glm::vec3>& points() const {return _points;}
+    const std::vector<glm::vec3>& normals() const {return _normals;}
 
 private:
     std::vector<Triangle> _triangles;
-    std::vector<glm::dvec3> _points;
-    std::vector<glm::dvec3> _normals;
+    std::vector<glm::vec3> _points;
+    std::vector<glm::vec3> _normals;
 };
 
 #endif // MESH_H
